Add an end-to-end test for custom objects
