+++ spec/graphql/authorization_spec.rb
@@ -954,66 +954,4 @@ describe "GraphQL::Authorization" do
       assert_equal expected_log, log
     end
   end
-
-
-  describe "overriding wrap" do
-    class WrapOverrideSchema < GraphQL::Schema
-      module LogTrace
-        def trace(key, data)
-          if ((q = data[:query]) && (c = q.context))
-            c[:log] << key
-          end
-          yield
-        end
-        ["parse", "lex", "validate",
-        "analyze_query", "analyze_multiplex",
-        "execute_query", "execute_multiplex",
-        "execute_field", "execute_field_lazy",
-        "authorized", "authorized_lazy",
-        "resolve_type", "resolve_type_lazy",
-        "execute_query_lazy"].each do |method_name|
-          define_method(method_name) do |**data, &block|
-            trace(method_name, data, &block)
-          end
-        end
-      end
-
-      module CustomIntrospection
-        class DynamicFields < GraphQL::Introspection::DynamicFields
-          def self.wrap(obj, ctx)
-            new(obj, ctx)
-          end
-        end
-      end
-
-      class Query < GraphQL::Schema::Object
-        def self.wrap(obj, ctx)
-          new(obj, ctx)
-        end
-        field :int, Integer, null: false
-        def int; 1; end
-      end
-
-      query(Query)
-      introspection(CustomIntrospection)
-      trace_with(LogTrace)
-    end
-
-    it "avoids calls to Object.authorized?" do
-      log = []
-      res = WrapOverrideSchema.execute("{ __typename int }", context: { log: log })
-      assert_equal "Query", res["data"]["__typename"]
-      assert_equal 1, res["data"]["int"]
-      expected_log = [
-        "validate",
-        "analyze_query",
-        "execute_query",
-        "execute_field",
-        "execute_field",
-        "execute_query_lazy"
-      ]
-
-      assert_equal expected_log, log
-    end
-  end
 end
